✨ Budget buttons for CRM Clients
